backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
openai==1.58.1
python-jose[cryptography]==3.3.0
